requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
python-jose
pytest
httpx
bcrypt
